Added missing fields to ifaddrs struct on Android
The original Android Open Source Project implementation of ifaddrs did not include the fields ifa_ifu and ifa_data. These field are present on the Rust side declaration of the structure ifaddrs.

The missing fields were added and are always set to NULL. The Rust side already deals with them being optional.