CBC mode with random IV, HMAC and DTLS Padding
Fixed the crypto CBC mode.

The Dtls Padding schema is used instead of Pkcs7.
IV is part of the payload, it will be read from the payload on decryption, and will be randomly generated and appended to the payload on encryption.
MAC on encryption was already present. On decryption it was being ignored. The MAC is now recalculated and compared with the received value using constant time operations.
MAC failure returns the same error as an encryption error in order to not leak sensitive information for potential attackers.