data_channel: allow acceptance of existing channels (#14)
(implements pion commit pion/datachannel@9710006. An associated PR for the webrtc repo will follow)

This allows negotiated channels to be fetched when created out of band of SDP by providing SCTP with a list of already created channels. This also includes a fix to Dial which will return a stream if already exists under the given stream id.